+++ airflow/ui/src/pages/Variables/ManageVariable/AddVariableModal.tsx
