+++ airflow/ui/src/pages/Variables/ManageVariable/DeleteVariableButton.tsx
