+++ airflow/ui/src/pages/Variables/ManageVariable/EditVariableModal.tsx
