+++ airflow/ui/src/pages/Variables/ManageVariable/VariableForm.tsx
